Update/easier csv download (#3)
* Updated to allow for easier download of .csv

In the previous version it was required to specify the delimeter for .csv. The goal of this update is to allow for downloading and reading .csv files without specifying the delimiter, if the delimiter is a `,`.

* Updated to allow for easier download of .csv

In the previous version it was required to specify the delimeter for .csv. The goal of this update is to allow for downloading and reading .csv files without specifying the delimiter, if the delimiter is a `,`.

* Clean up code

* Updated readme